console: Convert component into container
