+++ pkg/webui/locales/en.json
@@ -685,6 +685,30 @@
   "console.containers.pubsub-formats-select.index.warning": "Pub/Sub formats unavailable",
   "console.containers.pubsubs-table.index.format": "Format",
   "console.containers.pubsubs-table.index.host": "Server host",
+  "console.containers.user-data-form.add.adminLabel": "Grant this user admin status",
+  "console.containers.user-data-form.add.adminDescription": "Admin status enables overarching rights such as managing other users or modifying entities regardless of collaboration status",
+  "console.containers.user-data-form.add.userDescPlaceholder": "Description for my new user",
+  "console.containers.user-data-form.add.userDescDescription": "Optional user description; can also be used to save notes about the user",
+  "console.containers.user-data-form.add.userIdPlaceholder": "jane-doe",
+  "console.containers.user-data-form.add.userNamePlaceholder": "Jane Doe",
+  "console.containers.user-data-form.add.emailPlaceholder": "mail@example.com",
+  "console.containers.user-data-form.add.emailAddressDescription": "Primary email address used for logging in; this address is not publicly visible",
+  "console.containers.user-data-form.add.emailAddressValidation": "Treat email address as validated",
+  "console.containers.user-data-form.add.emailAddressValidationDescription": "Enable this option if you do not need this user to validate the email address",
+  "console.containers.user-data-form.edit.adminLabel": "Grant this user admin status",
+  "console.containers.user-data-form.edit.adminDescription": "Admin status enables overarching rights such as managing other users or modifying entities regardless of collaboration status",
+  "console.containers.user-data-form.edit.userDescPlaceholder": "Description for my new user",
+  "console.containers.user-data-form.edit.userDescDescription": "Optional user description; can also be used to save notes about the user",
+  "console.containers.user-data-form.edit.userIdPlaceholder": "jane-doe",
+  "console.containers.user-data-form.edit.userNamePlaceholder": "Jane Doe",
+  "console.containers.user-data-form.edit.emailPlaceholder": "mail@example.com",
+  "console.containers.user-data-form.edit.emailAddressDescription": "Primary email address used for logging in; this address is not publicly visible",
+  "console.containers.user-data-form.edit.emailAddressValidation": "Treat email address as validated",
+  "console.containers.user-data-form.edit.emailAddressValidationDescription": "Enable this option if you do not need this user to validate the email address",
+  "console.containers.user-data-form.edit.deleteTitle": "Are you sure you want to delete this account?",
+  "console.containers.user-data-form.edit.deleteWarning": "This will <strong>PERMANENTLY DELETE THIS ACCOUNT</strong> and <strong>LOCK THE USER ID AND EMAIL FOR RE-REGISTRATION</strong>. Associated entities (e.g. gateways, applications and end devices) owned by this user that do not have any other collaborators will become <strong>UNACCESSIBLE</strong> and it will <strong>NOT BE POSSIBLE TO REGISTER ENTITIES WITH THE SAME ID OR EUI's AGAIN</strong>. Make sure you assign new collaborators to such entities if you plan to continue using them.",
+  "console.containers.user-data-form.edit.purgeWarning": "This will <strong>PERMANENTLY DELETE THIS ACCOUNT</strong>. Associated entities (e.g. gateways, applications and end devices) owned by this user that do not have any other collaborators will become <strong>UNACCESSIBLE</strong> and it will <strong>NOT BE POSSIBLE TO REGISTER ENTITIES WITH THE SAME ID OR EUI's AGAIN</strong>. Make sure you assign new collaborators to such entities if you plan to continue using them.",
+  "console.containers.user-data-form.edit.deleteConfirmMessage": "Please type in this user's user ID to confirm.",
   "console.containers.user-data-form.edit.updateSuccess": "User updated",
   "console.containers.user-data-form.edit.deleteSuccess": "User deleted",
   "console.containers.users-table.index.invite": "Invite user",
